+++ serie2/lukecreated-serie2/src/primitives/navBar.jsx
@@ -1,0 +1,5 @@
+import { styled } from "../styles/stitches.config";
+
+const NavBar = styled("nav", {});
+
+export default NavBar;