+++ serie2/lukecreated-serie2/src/primitives/scaffold.jsx
@@ -1,0 +1,10 @@
+import { styled } from "../styles/stitches.config";
+
+const Scaffold = styled("div", {
+  //   color: "$main1",
+  //   backgroundColor: "$gray1",
+  padding: "1rem",
+  backgroundImage: "linear-gradient(var(--colors-gray3),var(--colors-main1) )",
+});
+
+export default Scaffold;