+++ changelog.md
@@ -68,3 +68,4 @@ Created on May 22, 2023
 ### Serie 2
 
 - Initial Development: Initialized Serie 2 development on July 30, 2024 -> __0.0.0__
+- Initial Development: Theme system is implemented on July 30, 2024 -> __0.0.1__